Add main file for a website
This is code that I used to deploy a static website with the tool openTofu. It is similar to the tool Terraform. 

This code will access openTofu and put the index.html file and the error.html file into an AWS S3 bucket. It will also change the bucket policy to public. 

This will make a website that anyone can access with a web browser.

This is a link to the website that is deployed with this code: 

http://website.it.handyman.s3-website-us-west-1.amazonaws.com/ 

.:[Note: you have to put your keys into the main.tf file for it to access your AWS account]:.